feat: Modified old UserController.cs file
